requirements.txt
flask
flask-restful
python-dotenv
sqlalchemy
psycopg2-binary